+++ test_files/types.d.ts
@@ -19,9 +19,4 @@ declare global {
   }
 }
 
-declare module "*.txt?raw" {
-  const content: string;
-  export default content;
-}
-
 export {};